feat: standalone mode
